requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pydantic-settings>=2.0.0
openai>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
